Make multiple agents per-build distinguishable
Using either host_name or service_instance_id is not sufficient.
They either fail when multiple nodes are used (same port) or multiple
parallel builds on one node (same hostname).
The solution is to combine the hostname with the port used for the
node_exporter. This should always be unique since only a single
node_exporter process can listen to this port at a time.