Support k8s container metrics
This is done by allowing the onMonit step to use custom dashboardUrls so
different pipelines can use different dashboards and by allowing to skip
the launch of the node_exporter and otelcol-contrib processes.
This supposes that metrics are gathered outside the control of this
plugin and are linked to the CI build.